refactor: split runtime lifecycle modules
